Adding LiteBag support and turning back on multi-addon check
LiteBag is now officially supported.  I've also added stubs for
cargBags_Nivaya and Inventorian, but it wasn't immediately obvious
how to support them, so I added a popup begging for help.  :)

I also re-enabled the check for multiple bag add-ons.  It now
properly accounts for whether or not you've disabled the all-in-one
bag in ElvUI in order to use a different bag.