Add support for Combuctor, Inventorian, cargBags, and fix WorldQuestTab
Combuctor and Inventorian are now fully supported.  cargBags works
great, too, but it requires a manual change until the author updates it
for you (I've put in the request).  The addon will inform you of the
change you need to make if it's missing for any reason.

Also, there was an issue with World Quest Tab buttons not getting
cleared out during scrolling.  This has been resolved.