Fixes for Encounter Journal and loot spec switching
When the Encounter Journal hadn't yet been opened, switching loot
spec would cause an error.  I fixed this and noticed that when
your toon is on auto loot spec, it wasn't correctly reflecting
status on Encounter Journal, so I fixed that.

I'm seeing some strange behavior where loot that should be valid
for a spec is (maybe) misidentified for that spec, but it's not
clear to me, yet, if I'm missing something, or if I'm actually
tracking it correctly and the game is wrong (seems unlikely).

At any rate, I don't think this is a new behavior, so I'm fixing
the crash and will go from there as I get more info.