Item Processing Updates
The Blizzard API doesn't always return item info which was resulting
in some calls getting hung up rather than completing.  Added a timeout
to catch this scenario and ensure we keep processing.  This may need
tweaked as it's more of a workaround, so I'll keep an eye on it.

Also fixed some errors that were happening in the Journal search and
added support for showing icons on the item sets!